+++ src/Kelly.Specs/Math/MatrixSpecs/MatrixHelper.cs
@@ -3,6 +3,6 @@ using Kelly.Math;
 
 namespace Kelly.Specs.Math.MatrixSpecs {
 	public class MatrixHelper {
-		public static readonly Matrix SequentialMatrix = new Matrix(Enumerable.Range(1, 16).Cast<float>());
+		public static readonly Matrix SequentialMatrix = new Matrix(Enumerable.Range(1, 16).Cast<double>());
 	}
 }
