+++ src/Kelly/Color.cs
@@ -1,13 +1,13 @@
 ﻿
 namespace Kelly {
 	public class Color {
-		public Color(float r, float g, float b) {
+		public Color(double r, double g, double b) {
 			R = r;
 			G = g;
 			B = b;
 		}
 
-		public float R, G, B;
+		public double R, G, B;
 
 		public static readonly Color Black = new Color(0, 0, 0);
 		public static readonly Color White = new Color(1, 1, 1);
