+++ src/Kelly/Math/Vector2.cs
@@ -3,24 +3,24 @@
 namespace Kelly.Math {
 	[DebuggerDisplay("({X}, {Y})")]
 	public class Vector2 {
-		public Vector2(float x, float y) {
+		public Vector2(double x, double y) {
 			X = x;
 			Y = y;
 		}
 
-		public float X { get; private set; }
-		public float Y { get; private set; }
+		public double X { get; private set; }
+		public double Y { get; private set; }
 
-		public static Vector2 operator *(Vector2 vector, float scalar) {
+		public static Vector2 operator *(Vector2 vector, double scalar) {
 			return new Vector2(vector.X * scalar, vector.Y * scalar);
 		}
 
-		public float SquaredLength {
+		public double SquaredLength {
 			get { return X * X + Y * Y; }
 		}
 
-		public float Length {
-			get { return (float)System.Math.Sqrt(SquaredLength); }
+		public double Length {
+			get { return System.Math.Sqrt(SquaredLength); }
 		}
 	}
 }
